Extracted Component interfaces and implementations into separate files with segregated tests.
