+++ src/display/stateful_test.go
